Fix: updateProfile의 deleteImage 이미지 값이 다를 때만 호출하기

+++ src/user/user.service.ts
@@ -136,14 +136,15 @@ export class UserService {
 			}
 		}
 
-		await this.imageService.deleteImage(user.image);
+		if (user.image && user.image !== image) {
+			await this.imageService.deleteImage(user.image);
+		}
 
 		user.username = username;
 		user.accountname = accountname;
 		user.intro = intro;
 		user.image = image;
 
-		console.log(user);
 		await user.save();
 
 		return {
